requirements/quality.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    make upgrade
#
amqp==5.2.0
    # via
    #   -r requirements/testing.txt
    #   kombu
annotated-types==0.7.0
    # via
    #   -r requirements/testing.txt
    #   pydantic
asgiref==3.8.1
    # via
    #   -r requirements/testing.txt
    #   django
astroid==3.2.4
    # via
    #   pylint
    #   pylint-celery
attrs==23.2.0
    # via
    #   -r requirements/testing.txt
    #   openedx-events
billiard==4.2.0
    # via
    #   -r requirements/testing.txt
    #   celery
camel-converter[pydantic]==4.0.1
    # via
    #   -r requirements/testing.txt
    #   meilisearch
celery==5.4.0
    # via
    #   -r requirements/testing.txt
    #   event-tracking
certifi==2024.7.4
    # via
    #   -r requirements/testing.txt
    #   elasticsearch
    #   requests
cffi==1.16.0
    # via
    #   -r requirements/testing.txt
    #   pynacl
charset-normalizer==3.4.0
    # via
    #   -r requirements/testing.txt
    #   requests
click==8.1.7
    # via
    #   -r requirements/testing.txt
    #   celery
    #   click-didyoumean
    #   click-log
    #   click-plugins
    #   click-repl
    #   code-annotations
    #   edx-django-utils
    #   edx-lint
click-didyoumean==0.3.1
    # via
    #   -r requirements/testing.txt
    #   celery
click-log==0.4.0
    # via edx-lint
click-plugins==1.1.1
    # via
    #   -r requirements/testing.txt
    #   celery
click-repl==0.3.0
    # via
    #   -r requirements/testing.txt
    #   celery
code-annotations==1.8.0
    # via
    #   -r requirements/testing.txt
    #   edx-lint
    #   edx-toggles
coverage[toml]==7.6.0
    # via
    #   -r requirements/quality.in
    #   -r requirements/testing.txt
    #   pytest-cov
ddt==1.3.1
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/testing.txt
dill==0.3.8
    # via pylint
django==4.2.14
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/testing.txt
    #   django-crum
    #   django-waffle
    #   edx-django-utils
    #   edx-toggles
    #   event-tracking
    #   openedx-events
django-crum==0.7.9
    # via
    #   -r requirements/testing.txt
    #   edx-django-utils
    #   edx-toggles
django-waffle==4.1.0
    # via
    #   -r requirements/testing.txt
    #   edx-django-utils
    #   edx-toggles
edx-ccx-keys==1.3.0
    # via
    #   -r requirements/testing.txt
    #   openedx-events
edx-django-utils==5.14.2
    # via
    #   -r requirements/testing.txt
    #   edx-toggles
    #   event-tracking
    #   openedx-events
edx-lint==5.3.7
    # via -r requirements/quality.in
edx-opaque-keys[django]==2.10.0
    # via
    #   -r requirements/testing.txt
    #   edx-ccx-keys
    #   openedx-events
edx-toggles==5.2.0
    # via
    #   -r requirements/testing.txt
    #   event-tracking
elasticsearch==7.13.4
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/testing.txt
event-tracking==2.4.0
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/testing.txt
fastavro==1.9.5
    # via
    #   -r requirements/testing.txt
    #   openedx-events
idna==3.10
    # via
    #   -r requirements/testing.txt
    #   requests
iniconfig==2.0.0
    # via
    #   -r requirements/testing.txt
    #   pytest
isort==5.13.2
    # via pylint
jinja2==3.1.4
    # via
    #   -r requirements/testing.txt
    #   code-annotations
kombu==5.3.7
    # via
    #   -r requirements/testing.txt
    #   celery
markupsafe==2.1.5
    # via
    #   -r requirements/testing.txt
    #   jinja2
mccabe==0.7.0
    # via pylint
meilisearch==0.31.5
    # via -r requirements/testing.txt
mock==5.1.0
    # via -r requirements/testing.txt
newrelic==9.12.0
    # via
    #   -r requirements/testing.txt
    #   edx-django-utils
openedx-events==9.11.0
    # via
    #   -r requirements/testing.txt
    #   event-tracking
packaging==24.1
    # via
    #   -r requirements/testing.txt
    #   pytest
pbr==6.0.0
    # via
    #   -r requirements/testing.txt
    #   stevedore
platformdirs==4.2.2
    # via pylint
pluggy==1.5.0
    # via
    #   -r requirements/testing.txt
    #   pytest
prompt-toolkit==3.0.47
    # via
    #   -r requirements/testing.txt
    #   click-repl
psutil==6.0.0
    # via
    #   -r requirements/testing.txt
    #   edx-django-utils
pycodestyle==2.12.0
    # via -r requirements/quality.in
pycparser==2.22
    # via
    #   -r requirements/testing.txt
    #   cffi
pydantic==2.9.2
    # via
    #   -r requirements/testing.txt
    #   camel-converter
pydantic-core==2.23.4
    # via
    #   -r requirements/testing.txt
    #   pydantic
pylint==3.2.6
    # via
    #   edx-lint
    #   pylint-celery
    #   pylint-django
    #   pylint-plugin-utils
pylint-celery==0.3
    # via edx-lint
pylint-django==2.5.5
    # via edx-lint
pylint-plugin-utils==0.8.2
    # via
    #   pylint-celery
    #   pylint-django
pymongo==3.13.0
    # via
    #   -r requirements/testing.txt
    #   edx-opaque-keys
    #   event-tracking
pynacl==1.5.0
    # via
    #   -r requirements/testing.txt
    #   edx-django-utils
pytest==8.3.1
    # via
    #   -r requirements/testing.txt
    #   pytest-cov
pytest-cov==5.0.0
    # via -r requirements/testing.txt
python-dateutil==2.9.0.post0
    # via
    #   -r requirements/testing.txt
    #   celery
python-slugify==8.0.4
    # via
    #   -r requirements/testing.txt
    #   code-annotations
pytz==2024.1
    # via
    #   -r requirements/testing.txt
    #   event-tracking
pyyaml==6.0.1
    # via
    #   -r requirements/testing.txt
    #   code-annotations
requests==2.32.3
    # via
    #   -r requirements/testing.txt
    #   meilisearch
six==1.16.0
    # via
    #   -r requirements/testing.txt
    #   edx-ccx-keys
    #   edx-lint
    #   event-tracking
    #   python-dateutil
sqlparse==0.5.1
    # via
    #   -r requirements/testing.txt
    #   django
stevedore==5.2.0
    # via
    #   -r requirements/testing.txt
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
text-unidecode==1.3
    # via
    #   -r requirements/testing.txt
    #   python-slugify
tomlkit==0.13.0
    # via pylint
typing-extensions==4.12.2
    # via
    #   -r requirements/testing.txt
    #   edx-opaque-keys
    #   pydantic
    #   pydantic-core
tzdata==2024.1
    # via
    #   -r requirements/testing.txt
    #   celery
urllib3==1.26.19
    # via
    #   -r requirements/testing.txt
    #   elasticsearch
    #   requests
vine==5.1.0
    # via
    #   -r requirements/testing.txt
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via
    #   -r requirements/testing.txt
    #   prompt-toolkit
